+++ pilz_trajectory_generation/CHANGELOG.rst
@@ -2,6 +2,12 @@
 Changelog for package pilz_trajectory_generation
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+0.3.4 (2019-02-05)
+------------------
+* refactor determining the trajectory alignment in the blend implementation
+* extend and refactor unittest of blender_transition_window
+* add planning group check to blender_transition_window
+
 0.3.3 (2019-01-25)
 ------------------
 * add more details to blend algorithm description